fix kl1m quotes schema

+++ hq.py
@@ -62,11 +62,11 @@ def kl1m_worker(q: Queue, year: int, output_dir: str):
                 "0": pl.Utf8,
                 "3": pl.Int32,
                 "4": pl.Int32,
-                "101": pl.Int64,
-                "102": pl.Int64,
-                "103": pl.Int64,
-                "104": pl.Int64,
-                "112": pl.Int64,
+                "101": pl.Int32,  # open, int64
+                "102": pl.Int32,  # high, int64
+                "103": pl.Int32,  # low, int64
+                "104": pl.Int32,  # last, int64
+                "112": pl.Int32,  # num_trades, int64
                 "113": pl.Int64,
                 "114": pl.Int64,
             },
